Resolves #210
The tests are taking advantage in some places of os.environ being a global namespace.  They check to make sure that KUBE_API_URL is set to "http://1.2.3.4", but they don't actually explicitly set it.  Plus, os.environ is a system call, so it really should be mocked, not set in the testcase itself.

Signed-off-by: Kevin McKenzie <kmckenzi@us.ibm.com>