Resolves issue #209 (#235)
This allows for indefinitely nested secrets files.  I also address #212, so if you override a secret loaded from a vault, it will be overridden the same way a secret loaded from the environment is overridden.  This is for consistency’s sake, but if there’s a reason to preserve the old behavior, that could be done.  I also made some changes to a couple of testcases to use the load_secrets function as opposed to the load_secrets_from_vault function.